+++ CHANGES.rst
@@ -4,6 +4,7 @@ Change log
 ** 2021-1-10 (2.0.0)**
 * Update to Ubuntu 20.04 / QGIS3 / python3 #67.
 * Simplify CLI.
+* Allow to process tasks for addresses only (-d).
 
 ** 2021-1-11 (1.9.3)**
 * Fix error deleting osm elements.